add support for grails-app/conf/simpleSpringSecurity.groovy
This is a first bit of support for configuring Spring security in a Grails 3 application.  This implementation is naive and overly simplified and is intended only as a starting point.  Right now what is supported is for an application to configure security with something like this (taken from http://github.com/grails-samples/grailstwitter)...

```
// grails-app/conf/simpleSpringSecurity.groovy
security = {
    authorize {
        authenticated "/status/**", "/updateStatus/**", "/users/**", "/unfollow/**", "/follow/**"
    }

    loginUrl '/login'
    logoutSuccessUrl '/'
    csrf false
}

users = {
    inMemory {
        newUser username: 'jeff', password: 'password', roles: ['USER']
        newUser username: 'graeme', password: 'password', roles: ['USER']
        newUser username: 'lari', password: 'password', roles: ['USER']
    }
}
```